driver/fingerprint/elan: Initialize errors to FP_ERROR_DEAD_PIXELS_UNKNOWN
The errors variable is currently initialized to 0 in fp_sensor_init. The
dead pixels count should be initialized to be UNKNOWN before the
maintenance routine is fired. Modify the initialization value of the
errors variable to FP_ERROR_DEAD_PIXELS_UNKNOWN, which includes both
initializing the bit fields corresponding to dead pixels to 1s and the
ones corresponding to the other error states to 0s.

BUG=b:379676720
TEST=CQ

Change-Id: I6592d1f6aa8e5cacf1e4fd568ac0835fc47ca8b1