driver/fingerprint/elan: Initialize errors to FP_ERROR_DEAD_PIXELS_UNKNOWN
The errors variable is currently initialized to zero in fp_sensor_init.
The dead pixels count should be initialized to be UNKNOWN before the
maintenance routine is fired. Modify the initialization value of the
errors variable to FP_ERROR_DEAD_PIXELS_UNKNOWN, which includes both
initializing the bit fields corresponding to dead pixels to ones and the
ones corresponding to the other error states to zeroes.

BUG=b:379676720
TEST=make BOARD=buccaneer -j100
     Flash firmware on Quincy Rev3 dev board
     Run fpinfo command from console
     Verify that error state is 0xC3FF

Change-Id: I6592d1f6aa8e5cacf1e4fd568ac0835fc47ca8b1