constitution: increase fanning by lowering temp_fan_max.
Chromebox-for-Meetings boards constitution and intrepid, used as
videoconferencing appliances, have a different duty cycle than ordinary
Chromebooks.  In the past, we wanted not to exceed 25 dB of fan noise
under ordinary loads; but because the fan-controlling temperature sensor
is not at the CPU, it can easily be 30 C cooler than the CPU, and having
a temp_fan_max of 90 C practically means that we'll never spin the fans
beyond 50% of its capable speed.

This trades noise budget for fan longevity (and hopefully meeting
quality with less thermal CPU throttling).

go/cfm-overheating-pt2
go/cfm-overheating

BUG=b:369348081
TEST=

Change-Id: Id50ce7414bdbe714bf00d64925d9ca7e9aa72c20