zephyr/program/ptlrvp: Update RTS54-PDC I2C addresses
Update the I2C addresses for the RTS54-PDC devices on the Intel RVP
platform to match the Realtek PDC FW version 1.50.1.

The previous addresses were applicable for PDC FW version 0.5.1,
causing the EC and PD controllers unable to communicate. This patch
updates the addresses to the correct values (0x66/0x67), allowing
the PD controllers to function properly.

BUG=b:376597467
TEST=Able to read the PDC FW version

```
24-11-20 20:41:01.318 Live: 1
24-11-20 20:41:01.319 FW Ver: 1.50.1
24-11-20 20:41:01.321 PD Rev: 769
24-11-20 20:41:01.321 PD Ver: 262
24-11-20 20:41:01.323 VID/PID: 18d1:5070
24-11-20 20:41:01.326 Running Flash Code: Y
24-11-20 20:41:01.328 Flash Bank: 0
24-11-20 20:41:01.329 Project Name: 'GOOG0400'
24-11-20 20:41:01.332 Driver Name: 'realtek_rts54_pdc'
24-11-20 20:41:01.334 FW Update: N
```

Change-Id: I6904e1130232f65f6ae8433679707db45c1270ec
Signed-off-by: Subrata Banik <subratabanik@google.com>