navi: Handle sleep hang for suspend workaround
This commit introduces a new hook, board_handle_sleep_hang(), enabling
boards to customize the handling of sleep hang events.

This change implements this hook on Navi to manage sleep hang events by
disabling the CS interrupt and configuring the S3 indicator GPIO as an
input.  This workarounds an issue where the system fails to suspend
correctly.

This change enhances the robustness of the sleep hang handling mechanism
by allowing boards to take specific actions in response to sleep hang
events.

BUG=b:354870788
TEST=sleeptimeout 100; Reproduce sleep hang and ensure the system can then
     successfully suspend.

Change-Id: I8c2dbc4ec1d323e213585ca2d10196c1869347eb