[TEST]: Initial Zephyr EC image
Patchset1
Create the initial Zephyr EC image for the new variant based on
the nereid reference board.

Patchset2: Remove support for sensors
It doesn't have any motions sensors. This CL removes support for
sensors from the initial project.

Patchset3: Remove support for keyboard
It doesn't have a keyboard. In addition, volume buttons are not
present on it, too.

BUG=none
TEST=zmake build pass

Change-Id: I1ba0132ad5003b9f48f2c3bb4d603e15b4580783