kconfig_check: switch to the Zephyr kconfiglib fork
Switch from the system wide to the Zephyr specific kconfiglib fork, this
is needed to support some extensions used upstream, specifically the
"configudefault" one.

This setup currently generates some warnings for symbols with no type
but it's otherwise good enough to fix downstreaming, will look into
fixing those as a followup.

search_paths is not supported ant not used, going to be removed in a
followup.

BUG=none
TEST=make -j5 BOARD=brya

Change-Id: If6497450b0514c8710dfb74ea90549b1aac56c4e