test: thermistor: scale for vref in the emulated read call
All Zephyr ADC APIs are going to be const for the time being, ADC being
changed in http://crrev.com/c/6060062. Since the internal vref is
implemented as a value in the api structure rather than a function call,
this means that it can't be mocked anymore, not without an api change
anyway.

Fix this in our test by accounting for that in the mocked ADC read
function instead.

BUG=none
TEST=./twister -v -s drivers/drivers.default.mock_power.ec_host_cmd

Change-Id: Iad050d2ca3d1b5270eff8044dd801258d81b83a8