usb_prl_sm: do not increment msgID if a message is discarded
This is a workaround to pass the msgID check in CTS COMMON.CHECK.PD.2.

The failure happens when message collision.

A message is queued in the protocol layer, but then immediately discards it
upon receiving a message.

MsgID is incremented according to 6.12.2.2.1.10, even if it is never
transmitted. Then the tester perceiving a message ID increment of 2, as it
never sees the discarded message.

To resolve this, the tx_discard_message state now only increments the msgID
if the message is successfully passed to the physical layer.

BUG=b:381999399, b:363100407
TEST=pass TEST.PD.EPR.SNK3.1

Change-Id: I569855a97948d2c8c8b8feaaa7c3edc603c4b184