usb_prl_sm: do not increment msgID if a message is discarded
This change workarounds a failure in CTS COMMON.CHECK.PD.2 caused by
message collisions.

The issue occurs when a message is queued in the protocol layer but
subsequently discarded upon receipt of another message. Although
discarded, the message ID is still incremented according to PD3.2
section 6.12.2.2.1.10. This leads to a perceived increment of 2 in the
message ID by the tester, as the discarded message is never transmitted.

To resolve this, the tx_discard_message state now only increments the
msgID if the message is successfully passed to the physical layer.

BUG=b:381999399, b:363100407 TEST=pass TEST.PD.EPR.SNK3.1

Change-Id: I569855a97948d2c8c8b8feaaa7c3edc603c4b184