pdc: Advertise 3A Src_Caps as a sink when available
Section 4.4 of the USB PD spec (Cable Type Detection) states that DRPs
acting as a sink must declare their full source capabilities when asked.
Sending Src_Caps as a sink is checked in COMMON.CHECK.PD.7#1 of the USB
PD CTS.

This updates the pdc power manager and dpm to advertise 3.0 A when
acting as a sink if it is currently available.

BUG=b:375665096
TEST=./twister -T zephyr/test/pdc, COMMON.CHECK.7 passes during
TEST.PD.PROT.SNK.2.

Change-Id: If08ca188d8b21125bc9a9912fcb9d998a55a17da