gsctool: Do not call GET_CHIP_ID unless unknown version
Cr50 doesn't support the GET_CHIP_ID TPMV command yet, and even though
gsctool handles that case, it still generates a UMA alert. Rely on the
major version number to determine device type first, and if that is out
of range fallback to the GET_CHIP_ID TPMV command. As of now, the
command should never get called.

BUG=b:376500403
TEST=BED=DT make tast TAST_EXPR=gscdevboard.GSCFactoryUpdate.ti50_0_21_1

Change-Id: Iedc0cb72c33d462302d2fa3e825b5902189c0356