pdc: tps6699x: Test changes for SET_SINK_PATH
This CL adds modifies the tps6699x emulator to set the interrupt status
bits sink_ready and new_power_cotract prior to emulating an
interrupt. These bits are required to be set so that task_srdy will be
completed without required a 1 second delay.

In addition, a callback function is added to the emulator for
task_aneg as the command SET_RDO will clear the flag
sink_enable_possible which needs to be set to avoid the one second
delay.

BUG=b:358274846
TEST=./twister -T zephyr/test/pdc

Change-Id: I8bfed6c861208548569eed74fc8a239d39db8ddd