case_closed_debugging_gsc: add note about chromeboxes
Chromeboxes turn off when the power button is pressed if no monitor is
connected. If someone wants to keep the AP up for the entire ccd open
procedure, they should connect a monitor to the DUT. Add a note about
this behavior in the press power button step of ccd open.

BUG=none
TEST=none

Change-Id: I3214c8058fe8a404cb74bf1ee75219a23fb85e4f