board/hyperdebug: Corrections to TPM "GSC ready" logic
When performing TPM operations via SPI or I2C, the GSC chips use an
extra signal to indicate when the host can continue the operation.  (For
read, it is generated after the address, to indicate when the requested
data is ready for reading, for writes, it is generated after the entire
transfer, to indicate when the GSC is ready to process the next
read/write.)

In order to reduce USB roundtrips, recently, code was added to
HyperDebug, to wait for such a high-to-low transition on a given pin, as
part of performing an I2C or SPI write or write-then-read operation.

That code suffers from a flaw, though.  In that the synchronous routines
to perform I2C or SPI may be slow to return, meaning that the leading
edge of the ready pulse has already taken place, by the time control
returns to HyperDebug code, which wants to look for the edge.  Strictly
speaking, it is not sufficient to look for a low level, since in theory
the low level could be the trailing part of a previous pulse.  However,
HyperDebug typically is not able to initiate successive operations with
less than 300us spacing, so the scenario of a previous pulse stretching
into the next TPM operation is not very realistic.

This patch inserts logic at the start of TPM operations to wait until
the ready signal has returned to its inactive high level (as noted
before, due to inherent overhead in HyperDebug USB communication, it is
unlikely that this code will ever encounter a trailing edge of a
previous pulse.)  This allows the code running after the actual SPI/I2C
transfer to merely look for a low level, rather than a falling edge,
which means that even if it is somewhat delayed, it will still detect
the pulse.

BUG=b:376508426
TEST=make tast TAST_EXPR=gscdevboard.GSCTPM*  (in ti50/common)

Change-Id: Icb306564144075ce8b5c5988e068f45f3ba1ab31