trulo: updates for GPIOs and I2C bus for the orb
This CL adds a few gpio signals that are present on the ORB, but are
not part of the Trulo reference design. In addition, the sensors i2c
bus was not being disabled and the ADC channel for temp_sensor_4 is
disabled to match the ORB.

The trulo reference includes 2 power button GPIOs. GPIO_00 and
GPIO_C7. The ORD is built with GPIO_00 connected to the power button
signal. ec_pwr_btn_odl (GPIO_OO) is correctly configured in
generated.dtsi and was then being deleted in gpio.dtsi and
gsc_ec_pwr_btn_odl was then being added which isn't required if
ec_pwr_btn_odl is used instead.

BUG=b:381019002
TEST=zmake build trulo/trulo-ti. Compared zephyr.dts between orisa and
trulo builds. Ran faft-ec and faft-pd tests.

Change-Id: I535f214889838a1e69d6af561518afd2179d08e8