board/hyperdebug: Allow clock adjustment
HyperDebug may be used to generate PWM waveforms, and in order to
achieve particular frequencies, it may be necessary for the system clock
to be a particular value.  Since users of HyperDebug may have various
requirements, a single system frequency will not satisfy everyone.  In
general, we want the highest frequency of 110MHz in order to get the
best possible performance of timer interrupts and edge detection
interrupts.

This patch allows a user to temporarily change the HyperDebug system
frequency to any value between 32MHz and 110MHz (in 2MHz increments).

BUG=b:384755415
TEST=Observe PWM waveform on oscilloscope (48MHz square)

Change-Id: Icd27dc2d781657c00db8d2fbf7cea319e350944f