chip/stm32: Rename STM32_HSE_CLOCK to CONFIG_STM32_CLOCK_HSE_HZ
The stm32h7 code used CONFIG_STM32_CLOCK_HSE_HZ while stm32l code used
STM32_HSE_CLOCK for the frequency of the external oscillator.

This patch mechanically renames the latter to the former, as I think it
makes more sense to have the CONFIG_ prefix, as this frequency depends
on external components on the board on which the EC sits, rather than
being some property of the STM family of chips.

BUG=b:384755415
TEST=make buildall_only

Change-Id: Id3961f5968ac642e513b899b8edb5fc069e4c214