board/hyperdebug: Fix off-by-one and rounding errors
It turns out, that if you want the PWM circuit to generate a square wave
that is high for two clock cycles and then low for two clock cycles,
then the timer ARR register must be set to 3 (one less than the total
period of four clock cycles), while the CCR register must be set to 2
(NOT one less than the high half-period).

Previously, the code erroneously subtracted one from the value of both
registers.

Also, several divisions rounded down, resulting in always using a
frequency as fast or faster than requested, when rounding to the nearest
possible value either slower or faster would be more appropriate.  This
is particularly important, as periods are given in (rounded) nanoseconds
rather than megahertz, meaning that e.g. a user asking for 32MHz (period
of 31.25ns) would result in opentitantool asking for a period of
31ns. Assuming a 96MHz HyperDebug base frequency, the period would be
2.976 clock cycles, which would be rounded down to a divisor of 2 rather
than 3, resulting in 48MHz output rather than 32MHz.

BUG=b:384755415
TEST=observe 48MHz square wave on oscilloscope with 50% duty cycle

Change-Id: Ib3be0a8fbed3320fdaac2cd5b60e5debc869a95c