board/hyperdebug: Add support for low power timers
The STM32L5 chip has three "low power" timer blocks, in addition to its
dozen or so "ordinary" timers.  We do not care about the ability for
those to run in low-power modes, but using them allows HyperDebug to
generate PWM waveforms on seven additional pins, including CN9.13, which
is the one used for feeding external clock to NT11 on the new shield.

Unfortunately, they seem to be a completely separate design, with
different registers.  The `chip/stm32` directory of the EC codebase does
not offer any drivers for them, so for now, the relevant logic is being
put into board/hyperdebug.

BUG=b:384755415
TEST=Observe 48MHz waveform on CN9.13 with oscilloscope

Change-Id: I6aa63f7dc45df37717e35269620a2ab3e8d71b27