DNS: Record Tx timestamp more accurately
In tcpci_tcpc_alert and other TCPC alert handlers, when the Tx-complete
bit is set in ALERT, pd_transmit_complete records the timestamp when
this was observed. This timestamp is used to calculate the ending time
for SenderResponseTimer.

Record this timestamp as early as possible in tcpci_tcpc_alert. Modify
helper functions to support this. Avoid extendng SenderResponseTimer
longer than intended. Modify other alert handlers to at least support
the new calling convention, if not improve timestamp accuracy.

Hack for event build.

BUG=b:361800172
TEST=twister -s zephyr/test/drivers/drivers.default

Change-Id: Iecacbd1e62ed3ef372133eff97fbc50b3f51b3dc