dsp_comms: Fix interrupt handling sequence
A few issues were found here:
1. During dsp_client_enable_interrupt() when we enable an edge-to-active
interrupt handler we probe the pin to see if it's already active. The
logic was flawed because the enable function was called for every
interrupt so we couldn't handle this edge case there. Instead, we now
handle it in the driver's init function and fake a one off interrupt.
2. In the interrupt handler we previously called k_event_post which
unblocked the CBI read thread. This works if we're trying to read a value
but fails if the EC just wants to send us a status. Instead, we should
submit the read_status_work and let the result from the status dictate
the call to k_event_post. This also resulted in a change to one of the
tests which depended on us NOT reading the status on the interrupt.

BUG=b:345530593
TEST=Flash Orisa and test tablet/clamshell mode

Change-Id: I39399ff772d7914c1e870dbc7d71ecef71be384c