orisa: Remove direct lid interrupt handling
Lid interrupt handling will be handled by DSP comms instead. Remove
references to the lid-open and tablet-mode interrupts from the ISH as
well as the notebook mode GPIO which will be re-purposed for the DSP
comms.

We need to temporarily disable tablet mode on the ISH until the DSP
comms can take that role.

BUG=b:345530593
TEST=Flash Orisa and test tablet/clamshell mode

Change-Id: I895ea1240cbc8f52960f9ea29bcd5ad645108fef