driver/fingerprint: sensor int pin check in fp_sensor_init
After sensor reset, the sensor will set the INT pin to the active level.
Once the sensor initialization is complete, the INT pin will be set to the non-active level.
By comparing the levels of the INT pin after the sensor reset and after initialization,
we can confirm whether the sensor's INT pin is properly connected to the host or not.

BUG=b:356211639
TEST=Cut the INT line on the FPC then restart EC. `Sensor IRQ not ready` will printed during `fp_sensor_init`.
And `Error State` will be 0x1000 when excute cmd `fpinfo`

Change-Id: I9c59630df7a1202a6fd8ff5ae9ef7eb857a55431