compare_builds.sh zephyr vpthon3 link fix
crrev/c/6048959 switched to the zephyr kconfig fork by adding a vpython3
file for a kconfig_check.py that linked to
../../../third_party/zephyr/main/vpython3

This broke compare_builds because compare builds checks out the ec
directory into a tmp directory, where the link to third_party is broken.

Because the link was broken, vpthyon3 dependencies were not met.
Specifically, kconfig_check.py broke when attempting to import pyyaml.

Fix is to add a link to the third_party directory.

BUG=b:215169275
TEST=./util/compare_builds --boards all

Change-Id: I9e99190801051a8db2e2d97531577854f0c33047