fatcat: Switch ECs to it8801 MFD driver
This updates the fatcat ITE and nuvoton add-in cards to use the it8801
MFD driver (from zephyr) instead of the cros-ec driver. This chip is
primarily used as an IO expander, so pull in the IO expander definitions
from the PTL RVP.

Note that the PTL RVP IO expander definitions also include the pca95xx
pin definitions. We could technically separate the it8801 and pca95xx
definitions, but that seems unnecessary at this point.

BUG=b:358178578,b:358181803
TEST=zmake build fatcat_npcx9m7f fatcat_it82002aw
	booted on fatcat_npcx9m7f and fatcat_it82002aw

Change-Id: Ibe582ddbbd37e2f57b1264b1372565d9a75261b9