trulo: Configure ec_wp_odl as ative low
This patch configures the ec_wp_odl GPIO with GPIO_ACTIVE_LOW, to
ensure the EC will interpret a low voltage on GPIOA pin 1 as the active
state for disabling the write protection.

Please note, the changes are currently WIP, as the generated.dsi should
be automatically generated by the ARB tool. This is currently under
process.

BUG=b:386741705
TEST=zmake build trulo

Change-Id: I7e0980f36dee850bb6a7d36a38fd54bb2192a3e4
Signed-off-by: Dinesh Gehlot <digehlot@google.com>