typec: Rename typec command to typec_legacy if naming conflict
This change addresses a potential naming conflict when both
EC-managed and PDC/AP-managed USB muxes are present, each with
their own `typec` console command.

When `CONFIG_PLATFORM_EC_CONSOLE_CMD_PDC_TYPEC` is enabled, the existing
`typec` command, which controls EC-managed muxes, is renamed to
`typec_legacy`. This ensures that the name `typec` is available for the
PDC/AP-managed mux command.

The Kconfig options are also updated to reflect this change and simplify
the configuration. The `PLATFORM_EC_CONSOLE_CMD_TYPEC_CHOICE` choice
block is removed, and the help text for `PLATFORM_EC_CONSOLE_CMD_TYPEC`
now explains the renaming behavior.

This change allows for a cleaner and more intuitive console experience
when both types of muxes are present in a system.

BUG=b:374014262 b:385253091
TEST=(1) check typec_legacy/typec both commands available
     (2) for existing project(s) (brox) that only uses
         PLATFORM_EC_CONSOLE_CMD_PDC_TYPEC,
         PLATFORM_EC_CONSOLE_CMD_TYPEC remains the same.

Change-Id: I37a7dab94c78f4be92874900f23e56a5c130ab22