trulo: Disable c++ malloc memory area
This CL disables the malloc feature for C++ support that's enabled for
the tokenizer. This prevents corruption of the panic memory area
that's allocated at the end of the unused RAM area.

Without this change, the malloc function assumes it can use all of
unallocated RAM and corrupts the panic memmory area which then leads
the EC getting stuck when a reboot happens after any panic instance.

BUG=b:388934636
TEST=verfied that EC reboot is successful following panics initiated
via the ec console crash command.

Change-Id: I6bb47cfb17207d9f86b7524ce5ce5077b5eb57f0