charge_state: prevent booting the AP when too hot/cold
is_battery_critical() will force a shutdown if the AP happens to be on
and the battery is too hot or cold to operate reliably or safely, which
could lead to awkward behavior where it might shut down due to
temperature then boot back up and shut down again. To prevent that,
extend `charge_prevent_power_on` to keep the AP from booting in the
first place if the battery is in a state that would cause an emergency
shutdown once it booted.

This should safely keep the AP off until the battery cools down if it's
too hot, and until it warms up or sufficient external power is available
when too cold. Whether the battery is allowed to charge is handled by
separate battery parameters, so this code only needs to consider the
rated temperatures for discharge.

Under very warm ambient conditions this could still lead to awkward
behavior when simply booting the AP causes the battery to exceed its
rated discharge temperature, but when power-on is prevented most systems
will show a distinctive pattern on the power or battery LEDs (such as
blinking red to indicate an error), so thermal cutout should at least
have some external indication even if it's not very clearly
communicated.

BUG=b:387931379
TEST=twister

Change-Id: If25b1a6aa10e2377eab2a635df3125e04a249d22