zephyr/test/fingerprint/task: Update expected values
System 'random()' returns number from 0 to INT_MAX, which means that
most significant bit was 0. On architectures with 64bit long it's even
worse because 4 most significant bytes were also 0.

Zephyr fixed that by taking only 3 least significant bytes [1].

Fingerprint tests for template encryption relies on the fact that the
"random" numbers are deterministic. This is generally true, but with
these type of changes we need to adjust some constants in our tests.

[1] https://github.com/zephyrproject-rtos/zephyr/pull/83446

BUG=None
TEST=Cherry-pick CL:6186370 to zephyr repository and run
     'twister -v -i -T zephyr/test/fingerprint'

Cq-Depend: chromium:6186370
Change-Id: I807642be472c752a938e499dbd52fdc17aaf84b1