pdc: Fix compiler warning
When building with clang it warns:

zephyr/subsys/pd_controller/pdc_console.c:222:52: error: result of
comparison of constant 255 with expression of type 'char' is always true
[-Werror,-Wtautological-constant-out-of-range-compare]
  222 |                           pdc_info.project_name[0] != 0xFF;
      |                           ~~~~~~~~~~~~~~~~~~~~~~~~ ^  ~~~~

When char is signed, it only has a range of -128 to 127. Cast it to a
uint8_t so the comparison to 255 is valid.

BUG=b:390722919
TEST=CQ

Change-Id: Ibe9e97b0313fce8df35b4eb94f7874366e03811f