pdc_rts54xx: Fix PDC_MAX_DATA_LENGTH
When building with clang it warns:

zephyr/drivers/usbc/pdc_rts54xx.c:1581:14: error: result of comparison
of constant 256 with expression of type 'uint8_t' (aka 'unsigned char')
is always true [-Werror,-Wtautological-constant-out-of-range-compare]
 1581 |                 assert(len <= ARRAY_SIZE(data->wr_buf));
      |                 ~~~~~~~~~~~^~~~~~~~~~~~~~~~~~~~~~~~~~~~

struct pdc_data_t has the following members:

/** PDC write buffer */
uint8_t wr_buf[PDC_MAX_DATA_LENGTH];
/** Length of bytes in the write buffer */
uint8_t wr_buf_len;
/** PDC read buffer */
uint8_t rd_buf[PDC_MAX_DATA_LENGTH];
/** Length of bytes in the read buffer */
uint8_t rd_buf_len;

The number of bytes in the wr_buf and rd_buf buffers is specified with
uint8_t, which has a maximum value of 255, so we need to change
PDC_MAX_DATA_LENGTH from 256 to 255 to allow the length to cover the
entire buffer.

Also switch from using ARRAY_SIZE(), which counts the number of elements
in the array, to sizeof(), which is the size of the array in bytes. This
future proofs the code if the size of the elements of the array changes.

BUG=b:390722919
TEST=CQ

Change-Id: I87db5bf39d0e0240755edbe36d70623a44917cc2