chip/stm32: Wait for core voltage to stabilize on STM32L4/5
When switching to high system clock frequencies, core voltage must be
increased first.  Previous code configured the voltage regulator with
the desired higher target range, but then immediately proceeded to
switch clock source.

Voltage rails take a long time to stabilise (compared to megahertz clock
frequencies), resulting the previous code occasionally misbehaving.

This patch insertes a loop to wait for the power regulator to report
that the new voltage is stable, before increasing the clock frequency.

(Furthermore, for STM32L4, the voltage selection logic was fatally
flawed in that it wrote to a wrong register.)

BUG=b:392973052
TEST=make BOARD=hyperdebug  (flash and observe no repeated rebooting)

Change-Id: Icc40e17ac037822d0cd38916021a95f81ea86675