pdc: Set and enforce driver initialization priorities
The PDC subsystem depends on the following parts that should initialize
in the following order, from first to last:

 * Charge Manager (89)
 * PDC device drivers (90)
 * pdc_power_mgmt (91)
 * PPM/UCSI (92)

The number in parenthesis is the new, default init priority for that
component (lower means earlier). All start in the POST_KERNEL phase
where application-supplied drivers are meant to go.

Add build assertions to enforce this ordering.

BUG=b:372910804
TEST=./twister -T zephyr/test/pdc

Change-Id: I5fb94cf6f9f26242fe03717351c76d73fa1c98ee