chip/stm32: Avoid stuck UART to USB forwarding
Some STM32 MCUs have a "UART RX timeout" interrupt, which we use for
improving USB transmission efficiency, in that we delay presenting data
on the USB endpoint, until we either have 64 bytes to fill a USB frame,
or the UART RX line drops idle for a little while.

It has been discovered, that if the UART RX line sits at a low level for
an extended perior (break condition), this will not trigger a timeout
intrrupt (as that apparently counts only when the RX line is at its
"idle" level which is high.)

If characters were received immediately before the break condition
started, they could previously be stuck until the break condition ends.
If the "break condition" is really the device on the other end powering
off, then that could be a long time, and it is highly unfortunate that
operators would not be able to see any final message printed on the
console before powering off.

When such a break condition starts, initially the UART will think that
it ist the beginning of a character frame, but when the stop bit does
not arrive, it will put a 0x00 into the RX FIFO, and at the same time
signal "framing error".  This patch adapts the code to notice such
framing errors, and in that case flush any queued characters, for
immediate transmission via USB.

BUG=b:393210880,b:378954948
TEST=Manually observe characters on /dev/ttyUSBn preceding a break

Change-Id: I48a456a2816cabd87a5fda173250cc9ecf584e8d