Add `attestation_ca.proto`.
Import the attestation CA's protobuf definitions. This will be used by
udsattestation to talk to the attestation CA.

These changes were tested by building them in a local Android.bp (which
will be added when these changes are merged into the AOSP copy of this
repository), and using the generated messages from udsattestation.

BUG=b:354037334

Change-Id: I40bbc0d2f78fbde660a535bc7f3036082dcad10b