pdc: initialize platform policy in pdc_subsys_init for DRP role
The PDC subsystem does not initialize the DRP role to PDC chips when EC
first boot. This leaves the connector status the same before any
platform policy applies. As a result, there is an edge case that when
the system is in DRP toggle on, and EC gets rebooted (such as reboot
from EC console, EC crashes and recovers, etc), the chipset is in off
status, then the DRP mode is not switch to force sink. In this case,
some peripherals are not going through a complete power cycle, and could
make them not working correctly.

Initialize the platform policy in `pdc_subsys_init` to the same policy
as the chipset is off, which sets the DRP role to force sink, so
peripherals get a complete power cycle. The DRP role will be updated by
other platform policies via power state change hooks.

BUG=b:386277572
TEST=Hook logic analyzer to CC pins with the port unattached, and
     (1) Ensure CC pins are toggling, then flash the EC. After reboot,
         the CC pins first stays in the force sink role, and starts
         toggling after chipset enters S0.
     (2) Ensure CC pins are toggling, then send EC reboot from EC
         console. The CC pins first stays in the force sink role, and
         starts toggling after chipset enters S0.
     (3) Ensure CC pins are toggling, then send `reboot hard ap-off`
         from EC console. The CC pins stays in the force sink role, and
         CC pins stars toggling only until the power button is pressed
         and chipset enters S0.

Change-Id: I9d7bf872aa7c722e7b7b5fed2551c581f1245e12