board/hyperdebug: Use external oscillator if present
The internal RC oscillator of HyperDebug is specified to be +/-1% at
25C.  If we ever want to verify that the GSC reacts to e.g. 3-second
button keypress within 1% of specified time, we will have to accept 2%
measured deviation, because HyperDebug could be 1% faster, and the GSC
1% slower, or vice versa.

For accurate time measurements, it would be useful to have more precise
clock in HyperDebug.  The MCU actually supports an external crystal or
other clock source.  This patch makes HyperDebug probe the external
source at startup, and if it is found, use it as reference for
generating its CPU clock frequency.  The external frequency if provided
must be a multiple of 4MHz.

Simply attaching a 16MHz crystal oscillator between CN11.29 and CN11.31
greatly improves the quality of the clock, but for optimal performance,
X3 should be populated with proper loading capacitors, and SB6/SB7
connected, while SB142/SB145 should be broken, to keep the clock traces
short.

BUG=none
TEST=Observe vastly reduced jitter on PWM from HyperDebug

Change-Id: Ic9df46672115fc924b31aff961af8f00417dfa75