board/hyperdebug: Use external oscillator if present
The internal RC oscillator of HyperDebug is specified to be +/-1% at
25C.  If we ever want to verify that the GSC reacts to e.g. 3-second
button keypress within 1% of specified time, we will have to accept 2%
measured deviation, because HyperDebug could be 1% faster, and the GSC
1% slower, or vice versa.

For accurate time measurements, it would be useful to have more precise
clock in HyperDebug.  The MCU actually supports an external crystal or
other clock source.  This patch makes HyperDebug probe the external
source at startup, and if it is found, use that for its CPU clock
frequency.

BUG=none
TEST=Observe vastly reduced jitter on PWM from HyperDebug

Change-Id: Ic9df46672115fc924b31aff961af8f00417dfa75