pdc: fix first boot not running init process in init_entry
The init_entry uses the send_cmd_return_state to ensure some init
processes only performed when transitioning into the PD_INIT state, and
is not just a state re-entry from the send command states. Since zephyr
will perform the zero-value initialization on declared variables, the
send_cmd_return_state is set to 0, which exactly matches to the enum
value PDC_INIT. As a result, the first boot init_entry will skip the
init process due to the current state is the same as the
send_cmd_return_state.

Add one more `PDC_INVALID` as the first enum value, so we can
distinguish it is the first time we enter the PD_INIT state and we
should perform initialization, and also explicitly initialize
the send_cmd_return_state.

BUG=none
TEST=build and flash to hylia device, the device can boot the system.

Change-Id: Ib079a65bd01b603b65b740e854ab8b8444c7a491