linker: Add noinit_end_of_ram section
Add noninit_end_of_ram memory section. It is placed at the very end of
ram where it is least likely to conflict with RO memory. It is only
used in the RW image.

Forcing a memory section to be at the end of ram in Zephyr is not well
supported. An extra check is added to the check_zephyr_end_of_ram script
to verify the section was placed as intended.

BUG=b:399207571
TEST=Build all

Change-Id: I5200bc1a46139b019b664567c84d2ba06d450628