common: Introduce preserved_ram_buf
Preserved ram buffers are semi persistent buffers that are intended to
persist across reboot and sysjumps. Integrity checks are builtin to so
so it is known when the buffer is valid or corrupted.

The buffers are designed to be low overhead and fast, especially for
writes. This makes them well suited for applications like logging and
tracing. Macros are used to generate the optimal code for a selected
configuration.

The implementation should be safe against memory faults caused by
corruption or out of bound offsets.

BUG=None
TEST=Build all

Change-Id: Ie757151baef6462a0bfd2005d5ffa80d9d879e5f