[WIP]panic: Introduce panic_log
Panic log captures a copy of the console log in uninitialized
preserved ram. After a panic, the panic_log is frozen until it is
dumped by the OS. Preserving the log is best effort. Integrity checks
ensure corruption is detected and the log is discarded.

The implementation is designed to be very low overhead.

BUG=None
TEST=Panic log preserved across watchdog and hang. Tested on lantis
and magolor.
TEST=Pass firmware.EcStress.all

Change-Id: I8797c5569f6d8bcc1df9655a2a6411c44f97eb90