chip/stm32: Refactor PLL setup for STM32L4/5
The STM32Lx chip can be clocked directly by an internal 16MHz
oscillator, or by an external clock, or by a PLL which can be based on
either of the two previous sources.

The logic in clock_init() was somewhat restricted, though.  In that the
mere declaration of the presence of external clock through
CONFIG_STM32_CLOCK_HSE_HZ was taken to imply that the PLL had to be
used, and had to be based on the external clock.  That is, it was not
possible to run the CPU directly on the external clock, nor was it
possible to set up such that a later command 'clock hse' could be used
to switch to external clock, without the external clock (and PLL) being
used as clock at boot.

The HyperDebug device could use an optional external clock, when
precision measurements are needed, and hence would need to define
CONFIG_STM32_CLOCK_HSE_HZ in order to be able to switch at runtime.
However, we want the initial clock of HyperDebug to be the PLL based on
the internal clock, in order for the device to also work in cases where
external clock is not present.

This patch rearranges the nesting of ifdefs, such that for STM32L4/5 the
define STM32_USE_PLL goes along with a new STM32_INITIAL_PLL_INPUT, and
together they control the boot-time clock setup.

There are only five boards in total which use STM32L4/5, and I have
verified that compare_build.sh reports identical binaries on those.

BUG=b:384755415
TEST=util/compare_build.sh --boards stm32l

Change-Id: I55168bfb2757662f12da19f1f1e6c5592cb652ad