npcx_monitor: Add a check for modifications
If the npcx_monitor binary changes, it needs to be updated in the servo
docker container.

Add a CMake script to check the sha256 of the output binary and fail the
build when detecting a mismatch.  The error message points the user to
check the README.md file for instructions on how to continue.

BUG=none
TEST=zmake build npcx_monitor - passes
TEST=zmake build npcx_monitor - fails on code change

Change-Id: I0cb3ce11e9fd497049c3a9ca696904bfc73941fb