pdc: rename platform policy to instructive names
Before, the PDC platform policy enforcing calls are confined in the
platform policy enforce worker, and is still under early development
stage, so the policy call name is not instructive. As we start to have
other places calling these platform policy, we need to rename these
policy calls with a readable name.

Rename these policy calls to the following 2 categories:
 - enforce_pd_chipset_<power state transition>_policy: the policy that
   is called when transitioning to another power state.
 - enforce_pd_chipset_<power state>_state_policy: the policy that is
   desired in a certain power state.

to make these calls readable and instructive.

BUG=none
TEST=./twister --testsuite-root zephyr/test/pdc/

Change-Id: I261957bb2bf72d1d534195c34639488ddbc580b2