chip/stm32: Support enabling watchdog in DFU RO stage
In order to allow recovering from a faulty RW image which locks up in
early initialization, it could make sense to enable the watchdog timer
in the minimal 4kB RO used in conjunction with ROM DFU for updating.
(This would be similar to how with "ordinary" EC RO/RW, the RO would
have enabled the watchdog long before jumping to the RW.

To support the required conditional compilation, a new makefile function
cfg_and has been introduced, following the pattern of cfg_not.

compare_build.sh reports that this patch does not change the resulting
binary for any stm32 boards.

BUG=none
TEST=util/compare_build.sh --boards stm32

Change-Id: I6069c161dc98c11f331aaf1878fa09e36ea83b44