board/hyperdebug: Enable watchdog in RO
A bad RW could potentially lock up during initialization on a fraction
of boards, due to e.g. process variation affecting the outcome of a race
condition not detected in testing.

With ordinary EC RO/RW the RO would have enabled hardware watchdog, so
that if the RW fails to boot, it can recover.  On the uServo-like USB
devices, which use DFU bootloader and minimal RO, the watchdog has until
now not been supported in the RO.

This patch enables a new feature in the DFU RO to enable watchdog before
transferring control to the RW.  The RO will count how many times
reboots have happened without the RW making a call to
dfu_bootmanager_clear(), and will eventually transfer control to the
STM32 DFU bootloader, rather than the RW, allowing recovery from a bad
push.

BUG=none
TEST=Observe DFU bootloader after intentinally upgrading to faulty RW image on HyperDebug

Change-Id: Ic0373ed99d5a588ced2bdca75b78a0c14a87672d