tps6699x: implement ucsi_set_pdos
This CL modifies the tps6699x driver to use UCSI_SET_PDOS instead of
the TI specific commands for setting either SRC or SNK PDOs. This
motivation for this change is that UCSI_SET_PDOS will automatically
trigger sending the updated SRC_CAPs to the port partner when a USB-PD
contract is already active. The TI specific command does not perform
this step.

BUG=b:352371324
TEST=Verfied via USB-PD analyzer that new SRC_CAPs are sent to the
port partner when updated when in a SRC power role. In addition,
verified that 3.0A can be drawn by the port partner.
Also verified './twister -T zephyr/test/pdc' passes

Change-Id: I6d51d65265449dc5529a8d79511890166bf339e3