npcx_monitor: Allow Zephyr toolchain hash as well
Add a switch on which hash to compare against based on the
ZEPHYR_TOOLCHAIN_VARIANT which will allow either Coreboot SDK or Zephyr
(currently using 0.17.0, so additional hashes might need to be added
later to support different version if needed). This is primarily
intended for local builds and tests comparing the 2 toolchains.

BUG=b:395569543
TEST=zmake build -t zephyr trulo

Change-Id: I9038c99d9a3424cc07dbf02c488b6090ed917803