[WIP]panic: Introduce panic_state
Panic state maintains the core state of the system in uninitialized
preserved ram. After a panic, the panic_state is frozen until it is
dumped by the OS. Preserving the panic_state is best effort. Integrity
checks ensure corruption is detected when it occurs.

The implementation is designed to be very low overhead.

BUG=None
TEST=Panic log preserved across watchdog and hang. Tested on lantis
and magolor.
TEST=Pass firmware.EcStress.all

Change-Id: I732873105eb2581cbcad45614c41534d821032b5