Create utility to check the cq excludes
We don't run most EC changes through the portage/ebuild based CQ. This
is generally a good thing, because the portage build is so slow, but
some of the ebuilds depend on random files in platform/ec.

Add a cq-excludes tool that will verify that the expected files are
included in the big portage based CQ runs.

BUG=b:393355553
TEST=Ran it

Change-Id: I2b756e2d208c945dd58648d9e867011cffb1c3b2