zephyr mchp: Add support for `wakeup` flag in VCI devicetree
Not all VCI input pins should be required to wake up the system from
hibernation. This CL enables `wakeup` flag to explicitly indicate that
a given VCI input is meant to be used as a wakeup source.

Although all VCI inputs are enabled, those without `wakeup` flag will
be disabled prior entering in hibernation.

BUG=b:396082465
TEST=Enable VCI input with `wakeup` flag
     Enter hibernation
     Use VCI input to wake up system
LOW_COVERAGE_REASON=Test to be added with b:399001492.

Change-Id: I9ee6e7d5b0d622ecb814f76c8a41d73853f091e5