test: pdc: Cover alternate mode connection as both sink and source
Test connecting an alternate mode port-partner by setting the attention
bit as both a source and a sink. This exercises both the source and sink
state machines.

Previously, we were only testing the case where the PDC was the source
(the sink case was covered accidentally through uninitialized change
bits; this has been corrected in chromium:6288068)

BUG=b:353713567
TEST=./twister -T zephyr/test/pdc --coverage

Change-Id: I9a8119c620839cdfa06037964b47adf8c94a9374