rauru: enable DP altmode event as wake-source
Enable the MKBP_EVENT_DP_ALT_MODE_ENTERED as a wake source for the
rauru platform. This allows the system to wake from S3 sleep when
a DP device is connected, triggering the altmode event.

This change modifies the `ec-mkbp-event-wakeup-mask` in the device
tree to include the DP altmode event in the wakeup mask.

BUG=b:396117735
TEST=In S3, plug-in DP plug to DP dongle at C0/C1 and wakes up

Change-Id: I3715ee9a98d5a5827071decf9a35e8157113b74f